refactor: relocate vue components to src and add header logic
